change views product
